% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjTable.R
\name{getAdjTable}
\alias{getAdjTable}
\title{Create an adjacency table for each cell}
\usage{
getAdjTable(rast, directions = 8)
}
\arguments{
\item{rast}{the raster being inspected}

\item{directions}{either 4, 8, 16, or 'bishop' to define adjacent/neighbors}
}
\value{
a table of ncell x nclr dimensions
}
\description{
Determine number of values present in every one of the raster's cells' neighborhood.
}
