% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeigh.R
\name{getNeigh}
\alias{getNeigh}
\title{Check for threshold}
\usage{
getNeigh(g, threshold = 3)
}
\arguments{
\item{g}{result from getAdjTable}

\item{threshold}{number of neighbors necessary to change cell's color}
}
\value{
a vector of integers representing colors of which most neighbors agree 0 represents no neighborhood agreement.
}
\description{
Function to see if enough neighbors create a threshold
}
