% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewVal.R
\name{getNewVal}
\alias{getNewVal}
\title{Get new Value for each cell}
\usage{
getNewVal(rast, neigh, nclr = 3)
}
\arguments{
\item{rast}{the raster being inspected}

\item{neigh}{results of getNeigh()}

\item{nclr}{number of colors used in simulation}
}
\value{
a vector of color integers representing colors
}
\description{
Determines the color of each cell from its neighbors
}
