% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initCells.R
\name{initCells}
\alias{initCells}
\title{Initialize cells}
\usage{
initCells(nrow, ncol, nclr = 3)
}
\arguments{
\item{nrow}{number of rows in raster}

\item{ncol}{number of columns in raster}

\item{nclr}{number of colors used in simulation}
}
\value{
a list of rasters representing each iteration of simulation
}
\description{
Create an initial raster to start simulation
}
