% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRPS.R
\name{plotRPS}
\alias{plotRPS}
\title{Animate results from RPS simulation}
\usage{
plotRPS(rastList, nclr, filePath)
}
\arguments{
\item{rastList}{list of rasters resulting from runRPS()}

\item{nclr}{number of colors used in runRPS()}

\item{filePath}{file name to provide gif}
}
\value{
Nothing is returned. A gif is created if successful
}
\description{
Create a .gif animation of RPS simulation
}
