% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRPS.R
\name{runRPS}
\alias{runRPS}
\title{Run Rock-Paper-Scissor Simulation}
\usage{
runRPS(
  nrow = 100,
  ncol = 100,
  nclr = 3,
  iterations = 100,
  directions = 8,
  threshold = 3
)
}
\arguments{
\item{nrow}{number of rows in raster}

\item{ncol}{number of columns in raster}

\item{nclr}{number of colors used in simulation}

\item{iterations}{number of iterations for simulation}

\item{directions}{either 4, 8, 16, or 'bishop' to define adjacent/neighbors}

\item{threshold}{number of agreeing neighbors necessary to change cell}
}
\value{
a list of rasters representing each iteration of simulation
}
\description{
Run RPS simulation over a set amount of iterations over a raster
}
